backend/requirements.txt
web3==4.0.0b5
Flask
Flask-Script
Flask-Restful
marshmallow
py-solc
